requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
pandas>=2.1.0
jupyter>=1.0.0
pytest>=7.4.0
Pillow>=10.0.0
tqdm>=4.66.0
python-dotenv>=1.0.0
